requirements.txt
faiss-cpu==1.7.2
torch>=1.7.0
torchvision==0.16.0
sentence-transformers==2.2.2
transformers>=4.11.3
numpy>=1.19.5
regex==2023.10.3
tqdm>=4.56.0
ollama>=0.1.5
langchain>=0.1.0
python-dotenv==1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
redis>=5.0.1
prometheus-client>=0.19.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
websockets>=12.0
yfinance>=0.2.35
googletrans>=3.1.0a0
python-multipart>=0.0.6
aiohttp>=3.9.0
loguru==0.7.2
scikit-learn>=0.24.1
pandas>=2.2.0
pandas-ta>=0.3.14b0
ydata-profiling>=4.6.3
tiktoken>=0.5.2
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
celery>=5.3.6
flower>=2.0.1
redis[hiredis]>=5.0.1
pillow==9.5.0
requests>=2.25.1
nltk>=3.6.1
rank-bm25>=0.2.2
flask>=2.0.1
flask-cors>=3.0.10
pyjwt>=2.1.0
matplotlib>=3.4.2
seaborn>=0.11.1
